Taskfile.yml: Add task to check test/ and negative_test/ git status
When running tests, it will create (or modify) a markdown file with
the command outputs. For now, the test are only checking that the commands
are exiting with correct status but not that the output is correct.

This has lead to:
- molecule.yml invalid syntax to go unnoticed
- tests for schema failures were marked as succeeding even if it was
  not due to missing check-jsonschema.

So add a task to check git status of test/ and negative_test/ and
use it inside the github CI. Don't do it by default after running
the tests since it may annoy the developers when adding new tests.

Signed-off-by: Arnaud Patard <apatard@hupstream.com>