meta: ensure author field is a valid github username
As v1 roles (standalone) used the author field as a github username
and all documentation examples uses it as a username, we require it
to match their requirements.

We also document the fact that `galaxy_info` field is not needed for
roles inside collections and that it can be removed.