meta: refactor schema versioning to improve errors (#385)
This greatly improves the quality of the validating error messages.

Improvement of error messages comes with a minor downside, as from
now on schema validation for meta would not pass on empty files. Users
will have to either put real content in meta files to pass schema
validation, such as version at least to pass. That was needed for
technical reasons as without it the error messages would be
considerably less easy to understand.

Related: https://github.com/python-jsonschema/jsonschema/issues/1002